tools/content_generator/requirements.txt
black
notion
numpy
python-slugify